refactor: split LearnPage into section components
