+++ apps/web/src/pages/learn/components/LiveArticle.tsx
@@ -1,0 +1,11 @@
+import { HowToPracticeArticle } from '../articles/HowToPracticeArticle'
+import { StubEntry } from './ArticleStub'
+
+export function LiveArticle({ id, title }: { id: string; title: string }) {
+  switch (id) {
+    case 'how-to-practice':
+      return <HowToPracticeArticle />
+    default:
+      return <StubEntry id={id} title={title} />
+  }
+}
